fix(cli): address PR review feedback on auto-peer-selection removal
Five fixes based on the PR review:

1. BUG: `isCompatibleBuyerProxy` in `buyer/start.ts` checked for the
   substring `'pinned peer'` which does NOT match the new 409 body
   containing `'No peer pinned'`. A second `antseed buyer start` would
   fail to detect the running proxy and crash with raw EADDRINUSE.
   Added explicit matches on `'no peer pinned'` and the JSON marker
   `'"no_peer_pinned"'`.

2. DESIGN: Change the "no peer pinned" response from 409 text/plain to
   400 application/json with the OpenAI/Anthropic-compatible error
   shape `{ error: { type, code, message, param, help } }`. SDKs now
   surface a proper `.message` on their error objects instead of a
   raw text/plain body. 400 matches "client request is missing
   required information" semantics that SDK retry logic handles as
   non-retryable.

3. MINOR: `onChainStatsRefreshedAt` was stamped with `Date.now()` in
   `network browse` even when no chain client was configured,
   producing a misleading "On-chain stats as of just now" header
   above a table with "—" in every on-chain column. Derive it from
   the peers' `onChainStatsFetchedAt` so it reflects only real RPC
   reads.

4. MINOR: `buildPaymentsConfig` was duplicated in `browse.ts` and
   `peer.ts`. Extracted to
   `apps/cli/src/cli/commands/network/chain-config-helper.ts` and
   shared.

5. NOTE: Added a caveat to the docblock on `resolveBestPaidPricing`
   calling out that cheapest input and cheapest output are picked
   independently across services, so the displayed pair may not
   correspond to any single real service offering. Point users to
   `--services` / `antseed network peer <id>` for exact comparison.
