+++ apps/cli/src/cli/commands/buyer/start.ts
@@ -162,10 +162,16 @@ async function isCompatibleBuyerProxy(port: number, timeoutMs = 1200): Promise<b
     if (antseedHeaderNames.some((header) => response.headers.has(header))) return true
 
     const body = (await response.text()).toLowerCase()
+    // Any of these markers indicate we reached an Antseed buyer proxy on this
+    // port. The `no_peer_pinned` case is the common one now that auto
+    // selection is disabled — a fresh proxy with no session pin answers
+    // /v1/models with a structured `{ error: { type: 'no_peer_pinned', ... } }`.
     return body.includes('no sellers available on the network')
       || body.includes('no peers support')
       || body.includes('p2p request failed')
       || body.includes('pinned peer')
+      || body.includes('no peer pinned')
+      || body.includes('"no_peer_pinned"')
   } catch {
     return false
   } finally {
